requirements.txt
fastapi~=0.111.0
pydantic~=2.7.2
redis~=5.0.4
hiredis~=2.3.2
uvicorn~=0.30.0
motor~=3.4.0
openai~=1.30.5
uvloop~=0.19.0
google-generativeai~=0.5.4
httpx~=0.27.0
aiohttp~=3.9.5
nextcord~=2.6.0
aiocache~=0.12.2
PyYAML~=6.0.1